+++ src/QNEthernet.h
@@ -27,7 +27,7 @@
 #include "qnethernet/compat/c++11_compat.h"
 #include "qnethernet/internal/optional.h"
 #include "qnethernet/lwip_driver.h"
-#include "qnethernet/security/RandomDevice.h"
+#include "qnethernet/security/random_device.h"
 #include "qnethernet/util/PrintUtils.h"
 #include "qnethernet_opts.h"